# Replace Offset/Trim with manual array slicing