… CliWrap/Internal/Extensions/CollectionExtensions.cs