Related to #83

Running  `HalfDuplexStreamBenchmarks` (new), with 1000000 random binary. Copied the benchmark logic from a unit test, more or less. Not sure it's the best benchmark for this scenario (it's only a 15 second run) but it got the job done in terms of measuring # allocations :-)

Before:
```
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100-rc.1.20452.10
  [Host]     : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
  DefaultJob : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT


|  Method |     Mean |   Error |  StdDev |   Median | Ratio |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|-------- |---------:|--------:|--------:|---------:|------:|----------:|----------:|---------:|----------:|
| CliWrap | 104.3 ms | 2.16 ms | 5.58 ms | 102.4 ms |  1.00 | 2800.0000 | 1000.0000 | 800.0000 |  14.25 MB |
```

After:
```
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100-rc.1.20452.10
  [Host]     : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
  DefaultJob : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT


|  Method |     Mean |   Error |  StdDev | Ratio |     Gen 0 |     Gen 1 |    Gen 2 | Allocated |
|-------- |---------:|--------:|--------:|------:|----------:|----------:|---------:|----------:|
| CliWrap | 113.5 ms | 2.25 ms | 3.70 ms |  1.00 | 1250.0000 | 1250.0000 | 750.0000 |   7.31 MB |
```